Improve parsing of date-times
* ParseDateTime retains its old implementation for now,
but is mostly deprecated and replaced by AsDateTime

* The big difference is that ParseDateTime parses each element
of the input vector, x, one at a time using lapply.  AsDateTime only
uses the first element of x to find a suitable order/format for parsing.
This means that AsDateTime will not work when the input vector x
contains elements with different formats/orders (the format of the first
element will be used to parse the entire vector).  This is consistent with
ParseDates and AsDate.

* The order in which different orders/formats are tested was changed
slightly from ParseDateTime, which sends (almost) all the orders to
lubridate::parse_date_time at once and lets lubridate figure things out.
This was done to pass all unit tests (e.g. "2 January 2016 00:34" was matching
"dbyHMS" instead of "dbYHM" because that was first in the old orders vector

* An argument exact (which gets passed to parse_date_time2)
was added to AsDateTime. exact = TRUE results in slightly faster
parsing, but may increase the chance of parsing failures. I experienced
this with a couple unit tests (e.g. "2 January 2016 00:34") until I changed
the order in which different orders were tested (as mentioned above)

* AsDateTime also has improved parsing of 'by' and 'bY'
formats, handling additional separators e.g. Jan-2012,
august13, aug/16, whereas ParseDateTime fails in those
cases.

Addresses: DS-1451