Check for ambiguous format in AsDate and AsDateTime
* AsDateTime now accepts a value of NULL for the us.format argument, and it is now
the default.  In this case, the function will check to see if the vector of dates can be parsed in both
dmXXX and mdXXX format.  If this is the case, a warning is thrown and the mdXXX (U.S.) format
is used.  Note that ParseDateTime has never allowed us.format to be NULL.

* AsDate also now allows for a NULL value of us.format, the check for ambiguity is more efficient
than what is done for ParseDates

* Additional support is provided for two-digit years in AsDateTime

* Both AsDate and AsDateTime now check to make sure there is no false positive with an order matching the first
element of the date vector, but not being correct for the rest of the dates; for example, mdY matching the first input,
but dmY clearly being correct for the entire vector.  Also relevant for distinguishing four digit years with no seconds
and two digit years with seconds (dmYHM or dmyHMS)

Addresses: DS-1451