Fix broken AsDate unit tests
* The first issue was that AsDate("9 or more") returns an NA of class
"POSIXct", "POSIXt" and not "logical", testthat::expect_equal complains of differing
classes even though the values are NA as expected, so the tests were updated to use
expect_true and is.na

* Could not track down why AsDate would not return NA of class logical in this case,
and it seems that the if statement in parsing.R:L264 is not entered

* An addtional unit test need to be fixed as
lubridate::parse_date_time2("More than 9", "by", exact = TRUE)
does not return NA for v1.6.0 at least on Windows.  From a practical standpoint,
if any other element of the vector of dates fails to parse, AsDate returns a vector
of NAs so that such false positives are unlikely to affect the user. hence, the unit
test was modified instead of modifying the AsDate to not check "by" format

Addresses: DS-1616